Add guard statement to ShowCursor call
In cases where mouse and keyboard events are occurring in rapid
succession, it is possible for the ShowCursor call to fire before a
HideCursor call has completed, leading to a BadMatch error.

In addition, it may be possible for multiple HideCursor calls to occur
before a ShowCursor call, leading to the cursor being stuck in a hidden
state.

I used a guard statement/loop to prevent premature ShowCursor calls and
call ShowCursor the correct number of times to mitigate both issues.