fix: cursor becomes hidden forever #26
If multiple keys are pressed and the mouse is moved simultaneously it
can cause the cursor to become hidden forever. This is because if
multiple keys are pressed very fast and the mouse moves at the same time
the X-Server generates respective events before xhidecursor is finished
processing the first event and changing the selected event-type. This
causes undefined states. To fix that synchronize the X-Server with
xhidecursor and discard all events from the event-queue after finishing
processing an event.